+++ src/hanami/src/core/processing/cpu/core_backpropagation.h
@@ -94,9 +94,6 @@ _backpropagateSection(SynapseSection* section,
     Axon* targetAxon = nullptr;
     constexpr float trainValue = 0.1f;
     float delta = 0.0f;
-    const float range = connection->potentialRange;
-    potential = range * static_cast<float>(potential > range)
-                + potential * static_cast<float>(potential <= range);
 
     // iterate over all synapses in the section
     while (pos < SYNAPSES_PER_SECTION && potential > POTENTIAL_BORDER) {