+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - updated core-structure for a cleaner communication between different hexagons
 - merge connection-, neuron- and synapse-block into one block object
 - updated section-processing for better input-handling especially for binary input data
+- replaced the old split-process for sections to avoid loosing half of the trained section in the process
 
 ### Changed